pyproject.toml
[project]
name = "swm-kkokkomu-video-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Guyeon Chung", email = "gouyeonch@naver.com" }
]
dependencies = [
    "google-cloud-speech>=2.26.1",
    "speechrecognition>=3.10.4",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
